Quick fix data type mismatch to run sample application.
+ TODO: Review fix.

+++ afreechart_sample/src/org/afree/chart/demo/DemoView.java
@@ -501,8 +501,8 @@ public class DemoView extends View
      *
      * @return The scale factor.
      */
-    public double getScaleX() {
-        return this.scaleX;
+    public float getScaleX() {
+        return (float) this.scaleX;
     }
 
     /**
@@ -511,8 +511,8 @@ public class DemoView extends View
      *
      * @return The scale factor.
      */
-    public double getScaleY() {
-        return this.scaleY;
+    public float getScaleY() {
+        return (float) this.scaleY;
     }
 
     /**
